Järjestä projektin perustietojen lomakkeen komponentteja fiksummin
